move cdm subset method under read
