+++ cognite/neat/_session/_read.py
@@ -28,7 +28,12 @@ from cognite.neat._issues.errors import NeatValueError
 from cognite.neat._issues.warnings import MissingCogniteClientWarning
 from cognite.neat._rules import catalog, importers
 from cognite.neat._rules.importers import BaseImporter
+from cognite.neat._rules.models.entities._single_value import ViewEntity
 from cognite.neat._rules.transformers import ClassicPrepareCore
+from cognite.neat._rules.transformers._converters import (
+    ToEnterpriseModel,
+    _SubsetEditableCDMRules,
+)
 from cognite.neat._utils.reader import NeatReader
 
 from ._state import SessionState
@@ -113,6 +118,86 @@ class CDFReadAPI(BaseReadAPI):
 
         importer = importers.DMSImporter.from_data_model_id(cast(NeatClient, self._state.client), data_model_id)
         return self._state.rule_import(importer)
+
+    def core_data_model(self, concepts: str | list[str]) -> IssueList:
+        """Subset the data model to the desired concepts.
+
+        Args:
+            concepts: The concepts to subset the data model to
+
+        Returns:
+            IssueList: A list of issues that occurred during the transformation.
+
+        Example:
+            Read the CogniteCore data model and reduce the data model to only the 'CogniteAsset' concept.
+            ```python
+            neat = NeatSession(CogniteClient())
+
+            neat.subset.data_model.core_data_model(concepts=["CogniteAsset", "CogniteEquipment"])
+            ```
+
+        !!! note "Bundle of actions"
+            This method is a helper method that bundles the following actions:
+            - Imports the latest version of Cognite's Core Data Model (CDM)
+            - Makes editable copy of the CDM concepts
+            - Subsets the copy to the desired concepts to desired set of concepts
+        """
+
+        concepts = concepts if isinstance(concepts, list | set) else [concepts]
+
+        self._state._raise_exception_if_condition_not_met(
+            "Subset Core Data Model",
+            empty_rules_store_required=True,
+            client_required=True,
+        )
+
+        warnings.filterwarnings("default")
+        ExperimentalFlags.core_data_model_subsetting.warn()
+
+        cdm_v1 = DataModelId.load(("cdf_cdm", "CogniteCore", "v1"))
+        importer: importers.DMSImporter = importers.DMSImporter.from_data_model_id(
+            cast(NeatClient, self._state.client), cdm_v1
+        )
+        issues = self._state.rule_import(importer)
+
+        if issues.has_errors:
+            return issues
+
+        cdm_rules = self._state.rule_store.last_verified_rules
+
+        issues.extend(
+            self._state.rule_transform(
+                ToEnterpriseModel(
+                    new_model_id=("my_space", "MyCDMSubset", "v1"),
+                    org_name="CopyOf",
+                    dummy_property="GUID",
+                    move_connections=True,
+                )
+            )
+        )
+
+        if issues.has_errors:
+            return issues
+
+        issues.extend(
+            self._state.rule_transform(
+                _SubsetEditableCDMRules(
+                    views={
+                        ViewEntity(
+                            space=cdm_v1.space,
+                            externalId=concept,
+                            version=cast(str, cdm_v1.version),
+                        )
+                        for concept in concepts
+                    }
+                )
+            )
+        )
+
+        if cdm_rules and not issues.has_errors:
+            self._state.last_reference = cdm_rules
+
+        return issues
 
     def graph(
         self,
